# backend/openapi/unital-v1.yaml
openapi: 3.1.0
info:
  title: Unital API
  version: 1.0.0
  description: |
    Building & complex management platform. Consolidated gateway API for the
    Unital Go microservices (identity, property, billing, facilities,
    operations, notifications/communications).

    **Auth**: `POST /auth/login` returns an access token (15 min) + refresh
    token. Send `Authorization: Bearer <access>` on every other call.

    **Errors**: all failures use RFC 9457 problem details with a
    machine-readable `code` and `trace_id`.
  contact: { name: Unital, url: https://unital.app }
servers:
  - url: http://localhost:8080/api/v1
    description: local gateway
  - url: http://localhost:18080/api/v1
    description: local gateway (dev port)
...
security:
  - bearerAuth: []
paths:
...
  /buildings/{buildingID}/charge-templates:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    post:
      tags: [Billing]
      summary: Define a recurring charge (manager/board)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChargeTemplateInput" }
      responses:
        "201": { description: Template created }
    get:
      tags: [Billing]
      summary: List charge templates (members)
      responses:
        "200": { description: Templates }
  /buildings/{buildingID}/charges:generate:
    post:
      tags: [Billing]
      summary: Generate charges for a period (idempotent per unit×template×period)
      parameters:
        - $ref: "#/components/parameters/BuildingID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [period]
              properties:
                period: { type: string, example: "1405-06", description: Jalali YYYY-MM }
                due_in_days: { type: integer, default: 14 }
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: integer }
                  period: { type: string }
        "403": { $ref: "#/components/responses/Problem" }
  /buildings/{buildingID}/charges:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    get:
      tags: [Billing]
      summary: List charges (filterable)
      parameters:
        - { name: unit_id, in: query, schema: { type: string } }
        - { name: period, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ChargeStatus" } }
      responses:
        "200":
          description: Charges
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Charge" }
  /charges/{chargeID}/payments:
    post:
      tags: [Billing]
      summary: Record a payment attempt (creates a pending transaction)
    parameters:
      - name: chargeID
        in: path
        required: true
        schema: { type: string }
    requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, method]
              properties:
                amount: { type: integer, format: int64, description: Rial }
                method: { type: string, enum: [online, cash, bank_transfer, cheque, pos] }
    responses:
        "201": { description: Transaction created }
  /transactions/{txID}/confirm:
    post:
      tags: [Billing]
      summary: Confirm a transaction (manager; applies to charge + invoice)
      parameters:
        - name: txID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Confirmed }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /buildings/{buildingID}/invoices:
    get:
      tags: [Billing]
      summary: Per-unit period invoices
      parameters:
        - $ref: "#/components/parameters/BuildingID"
        - { name: unit_id, in: query, schema: { type: string } }
      responses:
        "200": { description: Invoices }
  /buildings/{buildingID}/reports/financial:
    get:
      tags: [Billing]
      summary: Collections report for a period (manager/board)
      parameters:
        - $ref: "#/components/parameters/BuildingID"
        - { name: period, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                type: object
                properties:
                  period: { type: string }
                  units: { type: integer }
                  total_billed: { type: integer, format: int64 }
                  total_collected: { type: integer, format: int64 }
                  outstanding: { type: integer, format: int64 }
                  overdue_count: { type: integer }
                  collection_rate: { type: number }

  # ---------------- Notifications ----------------
...
  /charge-templates/{templateID}:
    patch:
      tags: [Billing]
      summary: Update a charge template (manager/board)
      parameters:
        - name: templateID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                amount: { type: integer }
                description: { type: string }
                is_active: { type: boolean }
      responses:
        "200": { description: Updated template }
        "403": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Billing]
      summary: Delete a charge template (manager/board)
      parameters:
        - name: templateID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
        "403": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BuildingID:
      name: buildingID
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with machine-readable code
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string, example: UNIT_NOT_IN_BUILDING }
        detail: { type: string }
        trace_id: { type: string }
...
    ChargeStatus:
      type: string
      enum: [pending, paid, partially_paid, overdue, cancelled]
    ChargeTemplateInput:
      type: object
      required: [name, amount]
      properties:
        name: { type: string, example: شارژ ماهانه }
        type:
          type: string
          enum: [monthly, maintenance, elevator, cleaning, security, green_space, pool, gym, other]
        amount: { type: integer, format: int64, description: Rial }
        description: { type: string }
    Charge:
      type: object
      properties:
        id: { type: string }
        unit_id: { type: string }
        template_id: { type: string }
        period: { type: string, example: "1405-06" }
        amount: { type: integer, format: int64 }
        paid: { type: integer, format: int64 }
        remaining: { type: integer, format: int64 }
        status: { $ref: "#/components/schemas/ChargeStatus" }
        due_date: { type: string, format: date }
...